+++ src/binding/cycles_CLEAN.ts
@@ -32,7 +32,6 @@ function expandGroupAnchors(edge: BindingEdge, source: BaseComponent, target: Ba
             return [...component.getAnchors().values()].map(a => a.id.anchorId);
         }
 
-        console.log('expanding group anchors for', component.id, anchorId, component,component.configurations)
         if(component.configurations[anchorId]) {
 
             const componentAnchors = [...component.getAnchors().values()]
@@ -204,10 +203,8 @@ function rewireMultiNodeConnections(
             return;
         }
 
-        console.log('creating internal anchor for', component.id, anchorId)
         // Create internal anchor
         createInternalAnchor(component, anchorId);
-        console.log('created internal anchor for2', component)
         
         // Redirect incoming edges to internal anchors
         redirectIncomingEdges(edges, nodeId, anchorId);
@@ -228,7 +225,6 @@ function rewireMultiNodeConnections(
             target: { nodeId, anchorId }
         });
 
-        console.log('pushed edges from ',nodeId, internalAnchorId, "to ", mergedNodeId, channel)
     });
     
     // Remove the original cycle edges
@@ -253,7 +249,6 @@ export function resolveCycleMulti(
     bindingManager: BindingManager
 ): BindingGraph {
     let processedGraph = cloneGraph(graph);
-    console.log('processedGraphONR!!!', processedGraph)
     // Keep resolving cycles until none are left
     let cycles = detectCyclesByChannel(processedGraph.edges);
     let count = 10;
@@ -565,7 +560,6 @@ function createInternalAnchor(component: BaseComponent, anchorId: string): void 
         return originalResult;
     };
 
-    console.log('about to set anchor', internalAnchorId, clonedAnchor)
     component.setAnchor(internalAnchorId, clonedAnchor);
 }
