+++ src/components/base.ts
@@ -57,7 +57,6 @@ export abstract class BaseComponent {
 
     bindings.forEach(({ value: binding, key }) => {
       const bindingProperty = key.startsWith('bind.') ? key.split('.')[1] : (key === 'bind' ? '_all' : key);
-      console.log('bindingProperty', bindingProperty)
 
       // TODO interactive binding reversal– this may be not needed depending on how scalar:scalar is handled
       if(bindingProperty === '_all'){
@@ -97,20 +96,8 @@ export abstract class BaseComponent {
         return 'span'
       }
 
-      // let configId = '';
-      //  if(bindingProperty.includes('bind')){
-      //   const split = bindingProperty.split('.')
-      //   const channel = split[0]
-      //   configId = split[1]
-      //   console.log('bindingProperty', bindingProperty, channel, configId)
-      //   // this.bindingManager.addBinding(this.id, , channel, configId);
-      // }
-      // console.log('configId', configId, 'binding',binding,'prop',bindingProperty)
-      // bindingProperty = configId;
-      
       // Check if this is a BaseChart by using instanceof or checking for chart-specific properties
       const isParentChart = ['Scatterplot','Histogram','BarChart'].includes(this.constructor.name);
-      console.log('isParentChart', isParentChart, this,this.constructor.name)
 
       if (isComponent(binding)) {
         this.bindingManager.addBinding(this.id, getTargetId(binding), bindingProperty, '_all');
@@ -120,7 +107,6 @@ export abstract class BaseComponent {
 
           const anchorSchema = Object.values(anchor.anchorSchema)[0];
           if(anchorSchema && anchorSchema.interactive && !isParentChart){ // TODO FIX such that chart isn't ddded...
-            console.log('dsadsaddads',this.id, this)
 
             this.bindingManager.addBinding(getTargetId(binding),this.id, anchor.id.anchorId, bindingProperty);
           }
@@ -130,7 +116,6 @@ export abstract class BaseComponent {
         // TODO: i think intertactiveity is not populating up and thus we don't get the inversee/internal stuff. 
         this.bindingManager.addBinding(this.id, getTargetId(binding), bindingProperty, binding.id.anchorId);
         if (binding.anchorSchema.interactive && !isParentChart) {
-          console.log('dsadsaddads2',this.id, this)
 
           this.bindingManager.addBinding(getTargetId(binding), this.id, binding.id.anchorId, binding.id.anchorId);
         }
